Add Easy Java Solution for 643. Maximum Average Subarray I from LeetCode

+++ LeetCode/Patterns/Sliding Window/643. Maximum Average Subarray I/Solution.java
@@ -1,21 +1,20 @@
 class Solution {
     public double findMaxAverage(int[] arr, int k) {
-         int sum = 0;
+        
+        int sum = 0;
 
-        // First window
         for(int i = 0; i < k; i++) {
             sum += arr[i];
         }
 
-        float max = sum;
+        int max = sum;
 
-        // Slide the window
         for(int i = k; i < arr.length; i++) {
             sum = sum + arr[i] - arr[i-k];
             max = Math.max(max, sum);
-
         }
-        float c = max/k;
+
+        double c = (double) max / k;
 
         return c;
     }
